[tagger] Remove all Fetch machinery
Since the introduction of the workloadmeta, tagger fetches are now gone
for good, and a `tagger.Tag` can no longer trigger side effects if an
entity does not exist. If an entity is not in the store, it's just not
in the store. Except for the garden collector, this is basically a
no-op. The garden collector is expected to move into the workloadmeta in
the near future as well.
