--- pkg/tagger/collectors/common.go
@@ -1,62 +1,0 @@
-// Unless explicitly stated otherwise all files in this repository are licensed
-// under the Apache License Version 2.0.
-// This product includes software developed at Datadog (https://www.datadoghq.com/).
-// Copyright 2016-present Datadog, Inc.
-
-package collectors
-
-import (
-	"encoding/json"
-	"strings"
-
-	"github.com/DataDog/datadog-agent/pkg/config"
-	"github.com/DataDog/datadog-agent/pkg/tagger/utils"
-	"github.com/DataDog/datadog-agent/pkg/util/log"
-)
-
-const (
-	// OrchestratorScopeEntityID defines the orchestrator scope entity ID
-	OrchestratorScopeEntityID = "internal://orchestrator-scope-entity-id"
-
-	autodiscoveryLabelTagsKey = "com.datadoghq.ad.tags"
-)
-
-// retrieveMappingFromConfig gets a stringmapstring config key and
-// lowercases all map keys to make envvar and yaml sources consistent
-func retrieveMappingFromConfig(configKey string) map[string]string {
-	labelsList := config.Datadog.GetStringMapString(configKey)
-	for label, value := range labelsList {
-		delete(labelsList, label)
-		labelsList[strings.ToLower(label)] = value
-	}
-
-	return labelsList
-}
-
-// mergeMaps merges two maps, in case of conflict the first argument is prioritized
-func mergeMaps(first, second map[string]string) map[string]string {
-	for k, v := range second {
-		if _, found := first[k]; !found {
-			first[k] = v
-		}
-	}
-
-	return first
-}
-
-func parseContainerADTagsLabels(tags *utils.TagList, labelValue string) {
-	tagNames := []string{}
-	err := json.Unmarshal([]byte(labelValue), &tagNames)
-	if err != nil {
-		log.Debugf("Cannot unmarshal AD tags: %s", err)
-	}
-	for _, tag := range tagNames {
-		tagParts := strings.Split(tag, ":")
-		// skip if tag is not in expected k:v format
-		if len(tagParts) != 2 {
-			log.Debugf("Tag '%s' is not in k:v format", tag)
-			continue
-		}
-		tags.AddHigh(tagParts[0], tagParts[1])
-	}
-}
