+++ pkg/tagger/collectors/types.go
@@ -20,7 +20,6 @@ type TagInfo struct {
 	LowCardTags          []string  // low cardinality tags safe for every pipeline
 	StandardTags         []string  // the discovered standard tags (env, version, service) for the entity
 	DeleteEntity         bool      // true if the entity is to be deleted from the store
-	CacheMiss            bool      // true if the TagInfo is generated by a tag miss
 	ExpiryDate           time.Time // keep in cache until expiryDate
 }
 
@@ -63,20 +62,13 @@ const (
 	HighCardinality
 )
 
-// Fetcher allows to fetch tags on-demand in case of cache miss
-type Fetcher interface {
-	Fetch(context.Context, string) ([]string, []string, []string, error)
-}
-
 // Streamer feeds back TagInfo when detecting changes
 type Streamer interface {
-	Fetcher
 	Stream() error
 	Stop() error
 }
 
 // Puller has to be triggered regularly
 type Puller interface {
-	Fetcher
 	Pull(context.Context) error
 }
