+++ pkg/tagger/collectors/workloadmeta_main.go
@@ -123,14 +123,6 @@ func (c *WorkloadMetaCollector) Stop() error {
 	return nil
 }
 
-// Fetch is a no-op in the WorkloadMetaCollector to prevent expensive and
-// race-condition prone forcing of pulls from upstream collectors.  Since
-// workloadmeta.Store will eventually own notifying all downstream consumers,
-// this codepath should never trigger anyway.
-func (c *WorkloadMetaCollector) Fetch(ctx context.Context, entity string) ([]string, []string, []string, error) {
-	return nil, nil, nil, nil
-}
-
 func workloadmetaFactory() Collector {
 	return &WorkloadMetaCollector{
 		store: workloadmeta.GetGlobalStore(),
@@ -169,6 +161,29 @@ func fargateStaticTags(ctx context.Context) map[string]string {
 	}
 
 	return tags
+}
+
+// retrieveMappingFromConfig gets a stringmapstring config key and
+// lowercases all map keys to make envvar and yaml sources consistent
+func retrieveMappingFromConfig(configKey string) map[string]string {
+	labelsList := config.Datadog.GetStringMapString(configKey)
+	for label, value := range labelsList {
+		delete(labelsList, label)
+		labelsList[strings.ToLower(label)] = value
+	}
+
+	return labelsList
+}
+
+// mergeMaps merges two maps, in case of conflict the first argument is prioritized
+func mergeMaps(first, second map[string]string) map[string]string {
+	for k, v := range second {
+		if _, found := first[k]; !found {
+			first[k] = v
+		}
+	}
+
+	return first
 }
 
 func init() {
