+++ pkg/tagger/local/tagger.go
@@ -15,18 +15,12 @@ import (
 	"github.com/DataDog/datadog-agent/pkg/util/log"
 
 	"github.com/DataDog/datadog-agent/cmd/agent/api/response"
-	"github.com/DataDog/datadog-agent/pkg/errors"
 	"github.com/DataDog/datadog-agent/pkg/status/health"
 	"github.com/DataDog/datadog-agent/pkg/tagger/collectors"
 	"github.com/DataDog/datadog-agent/pkg/tagger/tagstore"
 	"github.com/DataDog/datadog-agent/pkg/tagger/telemetry"
 	"github.com/DataDog/datadog-agent/pkg/tagger/types"
 	"github.com/DataDog/datadog-agent/pkg/util/retry"
-)
-
-const (
-	notFoundTTL = 5 * time.Minute
-	errTTL      = 30 * time.Second
 )
 
 // Tagger is the entry class for entity tagging. It holds collectors, memory store
@@ -37,7 +31,6 @@ type Tagger struct {
 	candidates map[string]collectors.CollectorFactory
 	pullers    map[string]collectors.Puller
 	streamers  map[string]collectors.Streamer
-	fetchers   map[string]fetcherEntry
 
 	store       *tagstore.TagStore
 	retryTicker *time.Ticker
@@ -52,11 +45,6 @@ type collectorReply struct {
 	instance collectors.Collector
 }
 
-type fetcherEntry struct {
-	fetcher   collectors.Fetcher
-	telemetry telemetry.FetcherTelemetry
-}
-
 // NewTagger returns an allocated tagger. You still have to run Init()
 // once the config package is ready.
 // You are probably looking for tagger.Tag() using the global instance
@@ -67,7 +55,6 @@ func NewTagger(catalog collectors.Catalog) *Tagger {
 		candidates: make(map[string]collectors.CollectorFactory),
 		pullers:    make(map[string]collectors.Puller),
 		streamers:  make(map[string]collectors.Streamer),
-		fetchers:   make(map[string]fetcherEntry),
 
 		store: tagstore.NewTagStore(),
 
@@ -211,7 +198,6 @@ func (t *Tagger) registerCollectors(replies []collectorReply) {
 			pull, ok := c.instance.(collectors.Puller)
 			if ok {
 				t.pullers[c.name] = pull
-				t.addFetcher(c.name, pull)
 			} else {
 				log.Errorf("error initializing collector %s: does not implement pull", c.name)
 			}
@@ -219,7 +205,6 @@ func (t *Tagger) registerCollectors(replies []collectorReply) {
 			stream, ok := c.instance.(collectors.Streamer)
 			if ok {
 				t.streamers[c.name] = stream
-				t.addFetcher(c.name, stream)
 				go stream.Stream() //nolint:errcheck
 			} else {
 				log.Errorf("error initializing collector %s: does not implement stream", c.name)
@@ -227,13 +212,6 @@ func (t *Tagger) registerCollectors(replies []collectorReply) {
 		}
 	}
 	t.Unlock()
-}
-
-func (t *Tagger) addFetcher(name string, collector collectors.Fetcher) {
-	t.fetchers[name] = fetcherEntry{
-		fetcher:   collector,
-		telemetry: telemetry.NewFetcherTelemetry(name),
-	}
 }
 
 func (t *Tagger) pull(ctx context.Context) {
@@ -260,69 +238,9 @@ func (t *Tagger) getTags(entity string, cardinality collectors.TagCardinality) (
 		return tagset.HashedTags{}, fmt.Errorf("empty entity ID")
 	}
 
-	cachedTags, sources := t.store.LookupHashed(entity, cardinality)
+	cachedTags := t.store.LookupHashed(entity, cardinality)
 
-	if len(sources) == len(t.fetchers) {
-		telemetry.QueriesByCardinality(cardinality).Success.Inc()
-		return cachedTags, nil
-	}
-
-	// Else, partial cache miss, query missing data
-	// TODO: get logging on that to make sure we should optimize
-
-	t.RLock()
-IterCollectors:
-	for name, collector := range t.fetchers {
-		for _, s := range sources {
-			if s == name {
-				continue IterCollectors // source was in cache, don't lookup again
-			}
-		}
-
-		log.Debugf("cache miss for %s, collecting tags for %s", name, entity)
-
-		var cacheMiss bool
-		var expiryDate time.Time
-		low, orch, high, err := collector.fetcher.Fetch(t.ctx, entity)
-		switch {
-		case errors.IsNotFound(err):
-			log.Debugf("entity %s not found in %s, skipping: %v", entity, name, err)
-			collector.telemetry.NotFound.Inc()
-
-			cacheMiss = true
-			expiryDate = time.Now().Add(notFoundTTL)
-		case err != nil:
-			log.Warnf("error collecting from %s: %s", name, err)
-			collector.telemetry.Error.Inc()
-
-			expiryDate = time.Now().Add(errTTL)
-		default:
-			collector.telemetry.Success.Inc()
-		}
-
-		// Submit to cache for next lookup
-		t.store.ProcessTagInfo([]*collectors.TagInfo{
-			{
-				Entity:               entity,
-				Source:               name,
-				LowCardTags:          low,
-				OrchestratorCardTags: orch,
-				HighCardTags:         high,
-				CacheMiss:            cacheMiss,
-				ExpiryDate:           expiryDate,
-			},
-		})
-	}
-	t.RUnlock()
-
-	cachedTags, _ = t.store.LookupHashed(entity, cardinality)
-
-	if cachedTags.Len() > 0 {
-		telemetry.QueriesByCardinality(cardinality).Success.Inc()
-	} else {
-		telemetry.QueriesByCardinality(cardinality).EmptyTags.Inc()
-	}
-
+	telemetry.QueriesByCardinality(cardinality).Success.Inc()
 	return cachedTags, nil
 }
