+++ pkg/tagger/replay/tagger.go
@@ -68,13 +68,13 @@ func (t *Tagger) Stop() error {
 
 // Tag returns tags for a given entity at the desired cardinality.
 func (t *Tagger) Tag(entityID string, cardinality collectors.TagCardinality) ([]string, error) {
-	tags, _ := t.store.Lookup(entityID, cardinality)
+	tags := t.store.Lookup(entityID, cardinality)
 	return tags, nil
 }
 
 // AccumulateTagsFor returns tags for a given entity at the desired cardinality.
 func (t *Tagger) AccumulateTagsFor(entityID string, cardinality collectors.TagCardinality, tb tagset.TagAccumulator) error {
-	tags, _ := t.store.LookupHashed(entityID, cardinality)
+	tags := t.store.LookupHashed(entityID, cardinality)
 
 	if tags.Len() == 0 {
 		telemetry.QueriesByCardinality(cardinality).EmptyTags.Inc()
